Update Copyright and License information for 2022 & Fixed copyright notices
