+++ README.md
@@ -1,4 +1,4 @@
-# OffSeason 2021
+# OffSeason 2021-2022
 #### BHS Devilbotz Team 2876
 
 ![Code Size](https://img.shields.io/github/languages/code-size/BHSDevilbotz/OffSeason2021?style=for-the-badge) ![Team](https://img.shields.io/badge/Team-2876-red?style=for-the-badge) ![Total Lines](https://img.shields.io/tokei/lines/github/BHSDevilbotz/OffSeason2021?style=for-the-badge) ![License](https://img.shields.io/badge/License-MIT%20%26%20FIRST%2FWPILib-green?style=for-the-badge) [![CodeQL](https://github.com/BHSRobotix/OffSeason2021/actions/workflows/codeql-analysis.yml/badge.svg?branch=master)](https://github.com/BHSRobotix/OffSeason2021/actions/workflows/codeql-analysis.yml)
