+++ src/main/java/bhs/devilbotz/commands/autonomous/drive/DriveTimed.java
@@ -1,7 +1,18 @@
-package frc.robot.commands.autonomous.drive;
+/*-------------------------------------------------------------------------------*/
+/* Copyright (c) 2021-2022 BHS Devilbotz. All Rights Reserved.                   */
+/* Open Source Software - may be modified, commercialized, distributed,          */
+/* sub-licensed and used for private use under the terms of the License.md       */
+/* file in the root of the source code tree.                                     */
+/*                                                                               */
+/* When doing any of the above, you MUST include the original                    */
+/* copyright and license files in any and all revised/modified code.             */
+/* You may NOT remove this header under any circumstance unless explicitly noted */
+/*-------------------------------------------------------------------------------*/
 
+package bhs.devilbotz.commands.autonomous.drive;
+
+import bhs.devilbotz.subsystems.DriveTrain;
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.DriveTrain;
 
 public class DriveTimed extends CommandBase {
     private final DriveTrain m_drive;
